ImageProcessing/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
certifi==2021.10.8
charset-normalizer==2.0.9; python_version >= '3'
idna==3.3; python_version >= '3'
pillow==8.4.0
requests==2.26.0
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
